// src/types/Notifications.ts
export enum NotificationType {
  USERS = 'Something went wrong!',
  POSTS = 'Unable to load posts',
  COMMENTS = 'Unable to load comments',
  ADD_COMMENT = 'Unable to add a comment',
  DELETE_COMMENT = 'Unable to delete a todo',
  // UPDATE_TODO = 'Unable to update a todo',
  // LOAD_TODOS = 'Unable to load todos',
}